requirements.txt
anthropic>=0.40.0
openai>=1.0.0
httpx>=0.27.0
pydantic>=2.0
streamlit>=1.37.0
python-dotenv>=1.0
pytest>=8.0
